+++ libs/ymir-core/include/ymir/hw/vdp/vdp_internal_callbacks.hpp
@@ -10,7 +10,7 @@
 namespace ymir::vdp {
 
 /// @brief Invoked when the HBlank and VBlank signals change.
-using CBHVBlankStateChange = util::RequiredCallback<void(bool level)>;
+using CBHVBlankStateChange = util::RequiredCallback<void(bool hb, bool vb)>;
 
 /// @brief Invoked when specific events occur while processing.
 using CBTriggerEvent = util::RequiredCallback<void()>;